Removes Git LFS because no GitHub Pages support
Replaces Git LFS tracking of images and fonts with regular git tracking using
the approach suggested in [1], because GitHub Pages does not support Git LFS
files (as of writing).

Removing Git LFS entirely was selected over updating URLs to absolute GitHub
paths (as done in [2]) as it's simpler to manage going forward.

[1] https://github.com/git-lfs/git-lfs/issues/3026#issuecomment-451598434 (not
    using the subsequent comment's '--renormalize' due to old Git version).

[2] https://github.com/git-lfs/git-lfs/issues/1342#issuecomment-333680934