Adds FAQ entry for individual partition operations
Adds an entry for the frequently asked questions discussing Rescuezilla's
limited ability to restore individual partitions. This has always been written
in the limitations page, but it's important to make this as clear as possible.

Providing retroactive support to the Sourceforge forum and seeing some very old
(circa 2017) questions about this has reminded me to improve the FAQ. Goes to
shows how valuable the retroactive support exercise is.