Adds more Patreon call-to-action links
Ensures there is a Patreon call-to-action orange bubble section to every single
page, top and bottom. Achieving sustainable funding is crucial to long-term
viability of Rescuezilla.

Also modifies the "completely free for personal and commercial use" to say
"freely available for personal and commercial use. All users should contribute
$1/month on the crowd funding website Patreon so that Rescuezilla can continue
to be developed", as this is more accurate to the GPLv3 license.