Shortens screenshot block by removing 2 images
Shortens the screenshot region on the about page, so it doesn't clip over the
into the next region by removing the screenshot of the Chromium web browser and
the screenshot of the desktop.

The raw images remain as they tend to be linked elsewhere on the internet by
myself but potentially others.