Clarifies that Rescuezilla is not a *trivial* fork
Clarifies that while Rescuezilla is indeed a fork, it's actually been rebuilt
from the ground up to support Clonezilla, and has features *beyond* Clonezilla.

This is to address misconceptions in earlier drafts of this blog post [1], as
well as in-person chats with people about Rescuezilla who understandably may
initially think it's a trivial fork of Redo Backup and Recovery without the
significant additional features and development of its complete rebuild from
the ground up.

[1] https://raywoodcockslatest.wordpress.com/2021/08/17/linux-v2p/